+++ example/src/App.tsx
@@ -72,7 +72,10 @@ export default class App extends React.Component<Props, State> {
   private onAppReady = async (): Promise<void> => {
     // this is for iOS only, Android TSAccountProtectionSDK is initialized from application onCreate.
     if (Platform.OS === 'ios') {
-      await TSAccountProtectionSDKModule.initializeIOS(config.clientId);
+      await TSAccountProtectionSDKModule.initializeIOS(config.clientId, null,{
+        enableTrackingBehavioralData: false,
+        enableLocationEvents: false
+      });
     }
 
     const isLogEnabled = true;